+++ charts/series.go
@@ -3,6 +3,7 @@ package charts
 import (
 	"github.com/go-echarts/go-echarts/v2/opts"
 	"github.com/go-echarts/go-echarts/v2/types"
+	"github.com/go-echarts/go-echarts/v2/util"
 )
 
 type SingleSeries struct {
@@ -527,7 +528,7 @@ func WithMarkPointNameCoordItemOpts(opt ...opts.MarkPointNameCoordItem) SeriesOp
 }
 
 func (s *SingleSeries) InitSeriesDefaultOpts(c BaseConfiguration) {
-	opts.SetDefaultValue(s)
+	util.SetDefaultValue(s)
 	// some special inherited options from BaseConfiguration
 	s.Animation = c.Animation
 }
