+++ components/page.go
@@ -17,7 +17,9 @@ const (
 type Charter interface {
 	Type() string
 	GetAssets() opts.Assets
+	// FillDefaultValues fill default values and would be override if any struct fields has been manually set
 	FillDefaultValues()
+	// Validate a validator and a post processor before render
 	Validate()
 }
