+++ util/default_val_setter.go
@@ -1,0 +1,42 @@
+package util
+
+import (
+	"reflect"
+	"strconv"
+)
+
+// SetDefaultValue set default values for the struct field.
+// inspired from: https://github.com/mcuadros/go-defaults
+func SetDefaultValue(ptr interface{}) {
+	elem := reflect.ValueOf(ptr).Elem()
+	walkField(elem)
+}
+
+func walkField(val reflect.Value) {
+	t := val.Type()
+
+	for i := 0; i < t.NumField(); i++ {
+		f := val.Field(i)
+		if f.Kind() == reflect.Struct {
+			walkField(f)
+		}
+
+		if defaultVal := t.Field(i).Tag.Get("default"); defaultVal != "" {
+			setField(val.Field(i), defaultVal)
+		}
+	}
+}
+
+// setField handles String/Bool types only.
+func setField(field reflect.Value, defaultVal string) {
+	switch field.Kind() {
+	case reflect.String:
+		if field.String() == "" {
+			field.Set(reflect.ValueOf(defaultVal).Convert(field.Type()))
+		}
+	case reflect.Bool:
+		if val, err := strconv.ParseBool(defaultVal); err == nil {
+			field.Set(reflect.ValueOf(val).Convert(field.Type()))
+		}
+	}
+}
